+++ app/schemas/project.py
@@ -30,7 +30,7 @@ class ProjectResponse(BaseModel):
     created_at: Optional[datetime] = None
     import_file_name: Optional[str] = None
     source_processing_status: Optional[str] = None
-    completeness_score:Optional[int]=None
+    completeness_score: Optional[float] = None
     data_quality_score:Optional[int]=None
     algorithm_used: Optional[str] = None 
     product_count:int=0
